pyproject.toml
[tool.poetry]
name = "nn_model"
version = "1.0.0"
description = ""
authors = [
    "Igor Adamski <igor.adamski2@gmail.com>",
]


[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.1"
numpy = "^1.24"
omegaconf = "^2.3.0"
loguru = "^0.7.2"
einops = "^0.7.0"


[tool.poetry.group.dev.dependencies]
black = "^21.7b0"
autoflake = "^1.4"
flake8 = "^3.9.2"
pytest = "^7.0.0"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
